requirements.txt
streamlit==1.37.1
pandas>=2.2
plotly>=5.22
numpy>=1.26
python-dateutil>=2.9
